moved contentcard to shared components
